+++ CHANGELOG.markdown
@@ -1,3 +1,7 @@
+Unreleased
+----------
+* Add tests/AdditionalTests.hs
+
 4.5.6 [2024.05.01]
 ------------------
 * Add specialized implementations of `log1p`, `expm1`, `log1pexp`, and